modularise weight frontend components
