+++ frontends/web/src/components/WeightComponents/WeightIndicator.css
