+++ frontends/web/src/components/WeightComponents/WeightIndicator.js
